Tag: move from Identity to util + use std lib defs
Also clang-format check-in.
